// src/services/models/commentResponse.ts
/**
 * Generated by orval v7.1.0 🍺
 * Do not edit manually.
 * A8 Pump Fun API
 * OpenAPI spec version: 0.0.12
 */
import type { UserResponse } from "./userResponse";

export interface CommentResponse {
  readonly content: string;
  readonly created_at: string;
  readonly id: number;
  readonly image: string;
  readonly isUserLike: boolean;
  readonly reply: number;
  readonly tokenAddress: string;
  readonly totalLike: number;
  readonly updated_at: string;
  readonly userInfo: UserResponse;
  readonly walletAddress: string;
}